+++ src/FakeXrmEasy.Plugins/PluginSteps/PluginStepRegistrationFieldNames/PluginStepRegistrationEntityNames.cs
@@ -1,0 +1,11 @@
+﻿namespace FakeXrmEasy.Plugins.PluginSteps.PluginStepRegistrationFieldNames
+{
+    internal class PluginStepRegistrationEntityNames
+    {
+        internal const string PluginType = "plugintype";
+        internal const string SdkMessageFilter = "sdkmessagefilter";
+        internal const string SdkMessage = "sdkmessage";
+        internal const string SdkMessageProcessingStep = "sdkmessageprocessingstep";
+        internal const string SdkMessageProcessingStepImage = "sdkmessageprocessingstepimage";
+    }
+}