+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [2.2.0]
 
+### Added 
+
+ - Added possibility to register plugin steps with logical name, and therefore support for late bound registration #38
+ - 
 ### Changed
 
 - Fix Sonar Quality Gate settings: DynamicsValue/fake-xrm-easy#28